refactor rename modsSlice -> modsPageSlice
