+++ src/extensionConsole/pages/mods/onboardingView/OnboardingView.tsx
@@ -23,7 +23,7 @@ import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faExternalLinkAlt } from "@fortawesome/free-solid-svg-icons";
 import marketplaceImage from "@img/marketplace.svg";
 import { type OnboardingType } from "@/extensionConsole/pages/mods/onboardingView/useOnboarding";
-import modsSlice from "@/extensionConsole/pages/mods/modsSlice";
+import modsPageSlice from "@/extensionConsole/pages/mods/modsPageSlice";
 import { useDispatch } from "react-redux";
 import workshopImage from "@img/workshop.svg";
 import { MODS_PAGE_TABS } from "@/extensionConsole/pages/mods/ModsPageSidebar";
@@ -49,7 +49,7 @@ const ActivateFromMarketplaceColumn: React.VoidFunctionComponent = () => (
 );
 
 const ActivateTeamBlueprintsColumn: React.VoidFunctionComponent = () => {
-  const { setActiveTab } = modsSlice.actions;
+  const { setActiveTab } = modsPageSlice.actions;
   const dispatch = useDispatch();
 
   return (
